[refactor] Create types from Context instance
This patch removes the public constructors for all types and makes them available on the Context class. Pointer, Array and Vector types are available as intoXType on the Type class. Ideally, you should create a Context whenever working with LLVM. All of the types are no longer retrievable without a context. If you do not wish to use a specified context, the global context should be used.