[refactor] Move primary constructors into Context (#204)
* [refactor] Merge Attribute into single sealed class

Previously, LLVM IR Attributes were spread into a llvm.ir.attributes package, containing an abstract class, two children and a common interface. This was a little overkill so this has now been reduced into a single class, exporting two typealiases.

* [refactor] Create types from Context instance

This patch removes the public constructors for all types and makes them available on the Context class. Pointer, Array and Vector types are available as intoXType on the Type class. Ideally, you should create a Context whenever working with LLVM. All of the types are no longer retrievable without a context. If you do not wish to use a specified context, the global context should be used.

* [refactor] replace constructs with member functions

* [tests] Update tests to use the new method over constructor apis

* [refactor] rename into*Type to get*Type

* [refactor] replace constructors with methods for constant values

* [tests] Update tests to use new constant constructors

* [style] format codebase