[rewrite] flatten class hierarchy and rewrite tests (#221)
At the time of writing this, the codebase is gigantic, split over 100+ files with deeply nested packages. This was a bad design decision and it should be fixed.

This PR moves and refactors most of the codebase under a single package, org.llvm4j.llvm4j, which is going to be the new project name.